refactor: split trace matrix helpers
